ENH: Add a script to plot the measure distribution
Add a script to plot the contrast-wise measure distribution.

Refactor the shared parts with the repeated measures ANOVA scripts to
common modules so that they can be reused.

Take advantage of the commit to ensure that the data frames do not
contain NaNs and inf values (which would impact negatively on the
assumption that the dependent variable should not contain outliers).