ENH: Allow axial and sagittal views in Slicer screenshot script
Allow axial and sagittal views in Slicer screenshot script.

Take advantage of the commit to improve the shell runner script to set
the background volume for the ground truth segmentation to the T1w
folder: previously it worked because that was the contrast whose
screenshots were being generated, and thus were not created again.