ENH: Add a grouped performance violin plot script
Add a grouped performance violin plot script: plots the performance data
of interest grouped according to the contrasts across the labels of
interest.

Add the accompanying utils methods to create the necessary labels for
the plots, etc.

Add the accompanying shell script to create the actual plots.

Add the `statannotations` package as a dependency.