ENH: Add labelmap volumetric screenshot scripts
Add scripts to create screenshots of a volumetric rendering of a
labelmap using VTK, including a script that add a slice of the
corresponding anatomical image.

Add the accompanying VTK helper functions.

Allow specifying the label of interest when extracting the polydata from
the image data so that the polydata corresponding to a given label can
be extract from the labelmap image data.