ENH: Adapt experiment scripts to process DWI subsampling data
Adapt experiment scripts to process DWI subsampling data.

Warn the user about missing participants in the subsampling experiment
performance computation. Since no check on the missing participant
ids is performed, the messages shown in the standard output should be
checked to ensure that the number of missing participants and ids
correspond to the expected numbers.

Use `np.hstack` instead of `np.array` and `ravel()` when aggregating
performance data across folds so that lists with different lengths can
be concatenated.