BUG: Handle appropriately optional labels argument in rmANOVA analysis
Handle appropriately optional labels argument in rm ANOVA analysis
script: only attempt to create the list of labels if these were
specified; use `None` otherwise.

The columns of interest in the rmANOVA data preparation function are
are optional. If `None`, all `SUIT` labels are used for the analysis.

Fixes:
```
  File "scripts/perform_anova_analysis.py", line 106, in main
    columns_of_interest=list(map(str, args.labels)),
TypeError: 'NoneType' object is not iterable
```