BUG: Create coefficient of variation output directory
Create coefficient of variation output directory: the output directory
basename depends on the input contrast, and it is appended to the output
directory rootname. This patch set adds the statement to effectively
create the output directory.

Oversight introduced in commit 6f0fefd.

Take advantage of the commit to change the output directory rootname to
the directory used effectively to store the data.