ENH: Add performance heatmap plot script
Add performance heatmap plot script: plots the mean performance values
read from a set of pandas description output files given as input as a
heatmap. Each row corresponds to a different file (contrasts in the case
of `dmriseg`), and each column is a group of labels.

Statistical significance data read from a file are plot according to a
binary code: significant values are marked with an "X". Only the `all`
label group is considered to provide significance information.